+++ test/compat/stateless.test.js
@@ -1,5 +1,5 @@
 /**
- * @overview Contains smoke tests for the main shescape module to verify
+ * @overview Contains smoke tests for the stateless shescape module to verify
  * compatibility with Node.js versions.
  * @license MIT
  */
@@ -8,33 +8,19 @@ import * as fc from "fast-check";
 
 import { arbitrary } from "./_.js";
 
-import { Shescape } from "../../index.js";
+import * as shescape from "../../stateless.js";
 
 export function testShescapeEscape() {
   fc.assert(
     fc.property(
-      arbitrary.shescapeOptions(),
       arbitrary.shescapeArg(),
-      (options, arg) => {
-        let shescape;
-
+      arbitrary.shescapeOptions(),
+      (arg, options) => {
         try {
-          shescape = new Shescape(options);
+          shescape.escape(arg, options);
         } catch (error) {
           const known = [
             "No executable could be found for ",
-            "Shescape does not support the shell ",
-          ];
-
-          if (!known.some((knownError) => error.message.includes(knownError))) {
-            throw new Error(`Unexpected error:\n${error}`);
-          }
-        }
-
-        try {
-          shescape.escape(arg);
-        } catch (error) {
-          const known = [
             "Cannot read property 'escape' of undefined",
             "Cannot read properties of undefined (reading 'escape')",
           ];
@@ -51,28 +37,14 @@ export function testShescapeEscape() {
 export function testShescapeEscapeAll() {
   fc.assert(
     fc.property(
-      arbitrary.shescapeOptions(),
       fc.array(arbitrary.shescapeArg()),
-      (options, args) => {
-        let shescape;
-
+      arbitrary.shescapeOptions(),
+      (args, options) => {
         try {
-          shescape = new Shescape(options);
+          shescape.escapeAll(args, options);
         } catch (error) {
           const known = [
             "No executable could be found for ",
-            "Shescape does not support the shell ",
-          ];
-
-          if (!known.some((knownError) => error.message.includes(knownError))) {
-            throw new Error(`Unexpected error:\n${error}`);
-          }
-        }
-
-        try {
-          shescape.escapeAll(args);
-        } catch (error) {
-          const known = [
             "Cannot read property 'escapeAll' of undefined",
             "Cannot read properties of undefined (reading 'escapeAll')",
           ];
@@ -89,28 +61,14 @@ export function testShescapeEscapeAll() {
 export function testShescapeQuote() {
   fc.assert(
     fc.property(
-      arbitrary.shescapeOptions(),
       arbitrary.shescapeArg(),
-      (options, arg) => {
-        let shescape;
-
+      arbitrary.shescapeOptions(),
+      (arg, options) => {
         try {
-          shescape = new Shescape(options);
+          shescape.quote(arg, options);
         } catch (error) {
           const known = [
             "No executable could be found for ",
-            "Shescape does not support the shell ",
-          ];
-
-          if (!known.some((knownError) => error.message.includes(knownError))) {
-            throw new Error(`Unexpected error:\n${error}`);
-          }
-        }
-
-        try {
-          shescape.quote(arg);
-        } catch (error) {
-          const known = [
             "Cannot read property 'quote' of undefined",
             "Cannot read properties of undefined (reading 'quote')",
             "Quoting is not supported when no shell is used",
@@ -128,28 +86,14 @@ export function testShescapeQuote() {
 export function testShescapeQuoteAll() {
   fc.assert(
     fc.property(
-      arbitrary.shescapeOptions(),
       fc.array(arbitrary.shescapeArg()),
-      (options, args) => {
-        let shescape;
-
+      arbitrary.shescapeOptions(),
+      (args, options) => {
         try {
-          shescape = new Shescape(options);
+          shescape.quoteAll(args, options);
         } catch (error) {
           const known = [
             "No executable could be found for ",
-            "Shescape does not support the shell ",
-          ];
-
-          if (!known.some((knownError) => error.message.includes(knownError))) {
-            throw new Error(`Unexpected error:\n${error}`);
-          }
-        }
-
-        try {
-          shescape.quoteAll(args);
-        } catch (error) {
-          const known = [
             "Cannot read property 'quoteAll' of undefined",
             "Cannot read properties of undefined (reading 'quoteAll')",
             "Quoting is not supported when no shell is used",